DEV-1144 split app dashboard to ssr
